requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.2.0
psycopg2==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
